requirements.txt
s3fs
python-dotenv
scipy
requests
